# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  docs.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
      success:
        type: boolean
    type: object
  docs.FoodDetailResponse:
    properties:
      allergen_info:
        type: string
      calories:
        type: number
      carbohydrates:
        type: number
      category:
        type: string
      fat:
        type: number
      fiber:
        type: number
      id:
        type: string
      image_url:
        type: string
      ingredients:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      nutrition_per_serving:
        additionalProperties:
          type: number
        type: object
      protein:
        type: number
      serving_size:
        type: string
      serving_size_unit:
        type: string
      sodium:
        type: number
      sugar:
        type: number
    type: object
  docs.FoodResponse:
    properties:
      calories:
        type: number
      carbohydrates:
        type: number
      category:
        type: string
      fat:
        type: number
      fiber:
        type: number
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      protein:
        type: number
      sodium:
        type: number
      sugar:
        type: number
    type: object
...
  docs.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  docs.RatingRequest:
    properties:
      comments:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    type: object
  docs.RatingResponse:
    properties:
      comments:
        type: string
      created_at:
        type: string
      food_id:
        type: string
      id:
        type: string
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  docs.Response:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nutrimatch.com
    name: API Support
    url: http://www.nutrimatch.com/support
  description: A personalized nutrition recommendation system built on the OpenNutrition
    dataset
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NutriMatch API
  version: "1.0"
paths:
...
  /api/v1/foods:
    get:
      consumes:
      - application/json
      description: Search for foods by name or other criteria
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/docs.FoodResponse'
                  type: array
                meta:
                  $ref: '#/definitions/docs.PaginationMeta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      summary: Search foods
      tags:
      - foods
  /api/v1/foods/{foodId}/rate:
    post:
      consumes:
      - application/json
      description: Add or update a user's rating for a food
      parameters:
      - description: Food ID
        in: path
        name: foodId
        required: true
        type: string
      - description: Rating information
        in: body
        name: rating
        required: true
        schema:
          $ref: '#/definitions/docs.RatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  $ref: '#/definitions/docs.RatingResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Rate a food
      tags:
      - foods
  /api/v1/foods/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific food
      parameters:
      - description: Food ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  $ref: '#/definitions/docs.FoodDetailResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      summary: Get food by ID
      tags:
      - foods
  /api/v1/foods/category/{category}:
    get:
      consumes:
      - application/json
      description: Get foods filtered by category
      parameters:
      - description: Food category
        in: path
        name: category
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                meta:
                  $ref: '#/definitions/docs.PaginationMeta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      summary: Get foods by category
      tags:
      - foods
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"